ci: rebalance integration shards by measured runtime, split game-flow
Why
---
Wall-clock crept from ~11m to 14m as new audio specs landed. Default
`--shard=N/3` partitions alphabetically by file, which put the heaviest
specs all on shard 1: 14m vs 4m vs 7m. Pulling per-spec durations from
run 25315290671's blob reports identified flow-12p-sheriff.spec.ts as
the dominant cost (~343s alone — 2× longer than game-flow's 13 tests
combined). Re-balancing without changing test code drops wall-clock to
~8m, free.

Changes
-------
1. Split `game-flow.spec.ts` (1461 lines, 13 tests). The bottom describe
   block ("Day 1 outcome scenarios") has 3 standalone tests that each
   `setupGame` independently — extracted into `game-flow-d1-outcomes
   .spec.ts` so the shard map can place them on a different runner.
   The 10 sequential STOMP-verification tests in the top describe block
   share `ctx` via `beforeAll` and stay together as before.
2. Replace `--shard=N/3` matrix with an explicit per-shard spec-list
   matrix (5 shards, hand-curated by measured runtime). Pin
   flow-12p-sheriff alone on shard 1 since it dominates wall-clock.
   Mapping documented in-line so the next person re-balancing knows
   the inputs.

Estimated wall-clock
--------------------
  Shard 1 (flow-12p-sheriff)             ~493s
  Shard 2 (game-flow + sheriff-flow)     ~410s
  Shard 3 (revote + idiot + d1-outcomes) ~400s
  Shard 4 (dead-role + werewolf + full)  ~360s
  Shard 5 (audio + login + guard)        ~265s

Wall-clock: max ≈ 8m (down from 14m). Each shard adds ~150s for backend
boot + Playwright/Gradle warmup, so the relative ordering is what
matters here.

Verification
  * vue-tsc --noEmit clean
  * Playwright list resolves 41 tests across 14 files (was 41 across 13)
  * Local run of shard 5 subset (login + audio-reconnect-recovery):
    5/5 pass in 1.3m, confirming positional spec-list invocation works
    end-to-end

Follow-ups (separate PRs)
  * #3 from CI-speedup table: workers > 1 within a shard. Requires
    fixing the host-login race in helpers/multi-browser.ts setupGame.
  * #4 from CI-speedup table: consolidate overlapping audio specs
    (bgm-flow / audio-reconnect-recovery / witch-audio / dead-role-audio
    / full-game-audio share 80% of setup).
  * Re-balance trigger: any spec that grows ≥30% in measured runtime
    relative to the values inlined in the workflow comment.
