+++ .github/workflows/ci.yml
@@ -182,14 +182,45 @@ jobs:
           retention-days: 7
 
   # ── E2E Integration (real backend) ──────────────────────────────────────────
+  #
+  # Shard map is hand-curated by measured spec runtime, not Playwright's
+  # default --shard=N/M (which alphabetically partitions and produced
+  # 14m/4m/7m wall-clock — flow-12p-sheriff alone took ~343s on shard 1).
+  # Re-balance whenever a new spec lands or runtime drifts ≥30%.
+  #
+  #   Last measurement (run 25315290671, commit 5a4a760, 2026-05-04):
+  #     flow-12p-sheriff.spec.ts        ~343s  (2 12p tests, biggest single)
+  #     revote-flow.spec.ts              ~98s
+  #     dead-role-audio.spec.ts          ~74s
+  #     idiot-flow.spec.ts               ~70s
+  #     full-game-audio.spec.ts          ~42s
+  #     game-flow.spec.ts (10 seq)      ~145s  (after split, was 13)
+  #     game-flow-d1-outcomes.spec.ts    ~80s  (split-out)
+  #     sheriff-flow.spec.ts            ~100s  (mostly setupGame; tests quick)
+  #     werewolf-win.spec.ts             ~91s
+  #     bgm-flow.spec.ts                 ~40s
+  #     audio-reconnect-recovery.spec.ts ~45s
+  #     witch-audio.spec.ts              ~80s
+  #     login.spec.ts                    ~10s
+  #     guard-audio-sequence.spec.ts      ~0s  (skipped)
   e2e-integration:
-    name: E2E · Integration (${{ matrix.shard }}/3)
+    name: E2E · Integration (${{ matrix.shard.id }}/5)
     runs-on: ubuntu-latest
     needs: backend
     strategy:
       fail-fast: false
       matrix:
-        shard: [1, 2, 3]
+        shard:
+          - id: 1
+            specs: e2e/real/flow-12p-sheriff.spec.ts
+          - id: 2
+            specs: e2e/real/game-flow.spec.ts e2e/real/sheriff-flow.spec.ts
+          - id: 3
+            specs: e2e/real/revote-flow.spec.ts e2e/real/idiot-flow.spec.ts e2e/real/game-flow-d1-outcomes.spec.ts
+          - id: 4
+            specs: e2e/real/dead-role-audio.spec.ts e2e/real/werewolf-win.spec.ts e2e/real/full-game-audio.spec.ts
+          - id: 5
+            specs: e2e/real/bgm-flow.spec.ts e2e/real/witch-audio.spec.ts e2e/real/audio-reconnect-recovery.spec.ts e2e/real/login.spec.ts e2e/real/guard-audio-sequence.spec.ts
 
     steps:
       - uses: actions/checkout@v4
@@ -237,15 +268,15 @@ jobs:
         run: npx playwright install-deps chromium
         working-directory: frontend
 
-      - name: E2E integration tests (shard ${{ matrix.shard }}/3)
-        run: npm run test:e2e:integration -- --shard=${{ matrix.shard }}/3
+      - name: E2E integration tests (shard ${{ matrix.shard.id }}/5)
+        run: npx playwright test --config=playwright.real.config.ts ${{ matrix.shard.specs }}
         working-directory: frontend
 
       - name: Upload blob report
         if: always()
         uses: actions/upload-artifact@v4
         with:
-          name: blob-report-integration-${{ matrix.shard }}
+          name: blob-report-integration-${{ matrix.shard.id }}
           path: frontend/blob-report/
           retention-days: 1
 
@@ -253,7 +284,7 @@ jobs:
         if: failure()
         uses: actions/upload-artifact@v4
         with:
-          name: playwright-report-integration-${{ matrix.shard }}
+          name: playwright-report-integration-${{ matrix.shard.id }}
           path: frontend/playwright-report/
           retention-days: 7
